+++ packages/neos-ui-ckeditor5-bindings/src/plugins/linkRelNofollow.js
@@ -1,27 +1,27 @@
 import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
 import LinkAttributeCommand from './linkAttributeCommand';
 
-const TARGET_BLANK = 'targetBlank';
+const REL_NOFOLLOW = 'linkRelNofollow';
 
-export default class TargetBlank extends Plugin {
+export default class RelNofollow extends Plugin {
     static get pluginName() {
-        return 'TargetBlank';
+        return 'LinkRelNofollow';
     }
     init() {
         const editor = this.editor;
-        editor.model.schema.extend('$text', {allowAttributes: TARGET_BLANK});
+        editor.model.schema.extend('$text', {allowAttributes: REL_NOFOLLOW});
         editor.conversion.attributeToElement({
-            model: TARGET_BLANK,
+            model: REL_NOFOLLOW,
             view: {
                 name: 'a',
                 attributes: {
-                    target: '_blank'
+                    rel: 'nofollow'
                 },
                 // the priority has got to be the same as here so the elements would get merged:
                 // https://github.com/ckeditor/ckeditor5-link/blob/20e96361014fd13bfb93620f5eb5f528e6b1fe6d/src/utils.js#L33
                 priority: 5
             }
         });
-        editor.commands.add(TARGET_BLANK, new LinkAttributeCommand(this.editor, TARGET_BLANK));
+        editor.commands.add(REL_NOFOLLOW, new LinkAttributeCommand(this.editor, REL_NOFOLLOW));
     }
 }
